requirements-dev.txt
pytest
pytest-mock
pytest-cov
pytest-pep8
tox
colorama
pyyaml
vcrpy
